requirements.txt
# Install PyTorch CPU version with: pip install torch --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
torchsummary>=1.5.1
wandb>=0.15.0
pyyaml>=6.0
pandas>=2.0.0
xgboost>=2.0.0

